Conflicts:
	src/SFA.DAS.EAS.Web/SFA.DAS.EAS.Web.csproj

+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Models/ApprenticeshipProvider/Provider.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Ukprn { get; set; }
+        public long Ukprn { get; set; }
         public string ProviderName { get; set; }
         public string Email { get; set; }
         public string Phone { get; set; }
